rename(NaverLogin): naverLogin > NaverLogin
